refactor: 기존의 Modal에 공통 컴포넌트 Modal 적용
